Makefile: add _GNU_SOURCE feature test macro
In order to use getopt() it is required to add the feature test macros
_POSIX_C_SOURCE >= 2 or _XOPEN_SOURCE. However, with these, getopt()
behaves as specified by POSIX and stops processing at the first
non-option argument. This behaviour is undesired. The GNU getopt(),
intead, does not stop in this situation and continues processing and
reordering the arguments, thus we use _GNU_SOURCE. Additionally,
it'll be necessary when implementing getopt_long().